+++ lib/src/service/entity/entity_b.dart
@@ -15,7 +15,7 @@ abstract class EntityB<T extends EntityI<T>> implements EntityI<T> {
 
   /// Entity database pointer.
   @override
-  int id = 0;
+  BigInt id = BigInt.from(0);
 
   /// Entity handling timestamp.
   @override