+++ lib/src/common/enums.dart
@@ -1,4 +1,4 @@
-/// [enum] for [Protocols].
+/// {enum} for [Protocols].
 ///
 /// Holds the available types of [Protocols] supported for communications.
 enum Protocols {
@@ -6,7 +6,7 @@ enum Protocols {
   http('http'),
   curl('curl');
 
-  /// Transaltion to a [core] implementations concept.
+  /// Transaltion to a {core} implementations concept.
   final String scheme;
 
   /// Creates a new [Protocols] instance.