+++ lib/src/service/decoding/decodable_i.dart
@@ -1,8 +1,12 @@
 import 'package:csm_client/csm_client.dart';
 
-/// [interface] for [Decodable] objects implementations, this interface provides
-/// members to generate an [TObject] based on an [Encoded] content.
+/// {interface} for [DecodableI].
+///
+/// Defines contract for an [Object] able to be loaded from [DataMap] object.
 abstract interface class DecodableI {
   /// Loads the current object content from the given [encode] object.
+  /// 
+  /// 
+  /// [encode] object that contains encoded content.
   void decode(DataMap encode);
 }