+++ CHANGELOG.md
@@ -2,9 +2,9 @@
 
 Changelog history for [csm_client] package.
 
-## CURRENT - [DD-MM-YYYY]
+## 8.0.1 - [04-05-2025]
 
-- Notes: N/A
+- Notes:
 
         1. [TracedException] the property [stackTrace] has gotten changes, now being optional at constructor and by default creating it from [StackTrace.current]. 
 
@@ -16,6 +16,20 @@ Changelog history for [csm_client] package.
         should throw an exception when the [key] is no present in the [DataMap].
 
         2. Reestructured how [get] type methods works internally.
+
+- Dependencies:
+
+    | Package                                 | Previous Version | Current Version |
+    |:----------------------------------------|:----------------:|:---------------:|
+    | http                                    | 1.3.0            | 1.3.0           |
+    | meta                                    | 1.16.0           | 1.16.0          |
+
+- Dev Dependencies:
+
+    | Package                                 | Previous Version | Current Version |
+    |:----------------------------------------|:----------------:|:---------------:|
+    | lints                                   | 5.1.1            | 5.1.1           |
+    | test                                    | 1.25.15          | 1.25.15         |
 
 ## 8.0.0 - [27.04-2025]